Updated to use NoireLib & fix issues supposedly
