--- Plugin/App/Struct/BlacklistedPlayer.cs
@@ -1,20 +1,0 @@
-using System;
-
-namespace PuppetMaster
-{
-    public class BlacklistedPlayer
-    {
-        public readonly string Id = "";
-        public string PlayerName = string.Empty;
-        public string PlayerWorld = string.Empty;
-        public bool Enabled = true;
-        public bool StrictPlayerName = true;
-
-        public BlacklistedPlayer(string name = "")
-        {
-            PlayerName = name;
-            string uniqueId = Guid.NewGuid().ToString();
-            Id = uniqueId;
-        }
-    }
-}
