+++ Plugin/Plugin.cs
@@ -3,82 +3,79 @@ using Dalamud.Interface.Windowing;
 using Dalamud.IoC;
 using Dalamud.Plugin;
 using Dalamud.Plugin.Services;
+using NoireLib;
 using System;
 using System.Collections.Generic;
 
-#nullable enable
-namespace PuppetMaster
+namespace PuppetMaster_Enhanced;
+
+public sealed class Plugin : IDalamudPlugin
 {
-    public sealed class Plugin : IDalamudPlugin
+    [PluginService] internal static IDalamudPluginInterface PluginInterface { get; private set; } = null!;
+
+    public string Name => "PuppetMaster - A.S Fork";
+
+    private readonly List<Tuple<string, string>> commandNames = new()
     {
-        [PluginService] internal static IDalamudPluginInterface PluginInterface { get; private set; } = null!;
+        new Tuple<string, string>("/puppetmaster", "Open PuppetMaster settings"),
+        new Tuple < string, string >("/ppm", "Alias for PuppetMaster settings"),
+        new Tuple < string, string >("/puppet", "Alias for PuppetMaster settings")
+    };
 
-        public string Name => "PuppetMaster - A.S Fork";
+    public WindowSystem WindowSystem = new("PuppetMaster - A.S Fork");
+    public ConfigWindow ConfigWindow { get; init; } = new();
 
-        private readonly List<Tuple<string, string>> commandNames = new()
+
+    public Plugin()
+    {
+        NoireLibMain.Initialize(PluginInterface, this);
+
+        Service.InitializeConfig();
+
+        WindowSystem.AddWindow(ConfigWindow);
+
+        foreach (var CommandName in commandNames)
         {
-            new Tuple<string, string>("/puppetmaster", "Open PuppetMaster settings"),
-            new Tuple < string, string >("/ppm", "Alias for PuppetMaster settings"),
-            new Tuple < string, string >("/puppet", "Alias for PuppetMaster settings")
-        };
-
-        public WindowSystem WindowSystem = new("PuppetMaster - A.S Fork");
-        public ConfigWindow ConfigWindow { get; init; } = new();
-
-
-        public Plugin()
-        {
-            PluginInterface.Create<Service>(Array.Empty<object>());
-
-            Service.InitializeConfig();
-
-            WindowSystem.AddWindow(ConfigWindow);
-
-            foreach (var CommandName in commandNames)
+            NoireService.CommandManager.AddHandler(CommandName.Item1, new CommandInfo(OnCommand)
             {
-                Service.CommandManager.AddHandler(CommandName.Item1, new CommandInfo(OnCommand)
-                {
-                    HelpMessage = CommandName.Item2
-                });
-            }
-
-            Service.ChatGui.ChatMessage += new IChatGui.OnMessageDelegate(ChatHandler.OnChatMessage);
-
-            PluginInterface.UiBuilder.Draw += DrawUI;
-            PluginInterface.UiBuilder.OpenConfigUi += DrawConfigUI;
-
-            Service.InitializeEmotes();
-            Service.InitializeWorlds();
+                HelpMessage = CommandName.Item2
+            });
         }
 
-        public void Dispose()
-        {
-            this.WindowSystem.RemoveAllWindows();
+        NoireService.ChatGui.ChatMessage += new IChatGui.OnMessageDelegate(ChatHandler.OnChatMessage);
 
-            Service.ChatGui.ChatMessage -= new IChatGui.OnMessageDelegate(ChatHandler.OnChatMessage);
+        PluginInterface.UiBuilder.Draw += DrawUI;
+        PluginInterface.UiBuilder.OpenConfigUi += DrawConfigUI;
+    }
 
-            foreach (var CommandName in commandNames)
-            {
-                Service.CommandManager.RemoveHandler(CommandName.Item1);
-            }
+    public void Dispose()
+    {
+        PluginInterface.UiBuilder.Draw -= DrawUI;
+        PluginInterface.UiBuilder.OpenConfigUi -= DrawConfigUI;
 
-            GC.SuppressFinalize(this);
-        }
+        NoireService.ChatGui.ChatMessage -= new IChatGui.OnMessageDelegate(ChatHandler.OnChatMessage);
 
-        private void OnCommand(string command, string args)
-        {
-            this.DrawConfigUI();
-        }
+        foreach (var CommandName in commandNames)
+            NoireService.CommandManager.RemoveHandler(CommandName.Item1);
 
-        private void DrawUI()
-        {
-            this.WindowSystem.Draw();
-        }
+        WindowSystem.RemoveAllWindows();
+        ConfigWindow.Dispose();
 
-        public void DrawConfigUI()
-        {
-            // Toggle open/close plugin
-            ConfigWindow.IsOpen = !ConfigWindow.IsOpen;
-        }
+        NoireLibMain.Dispose();
+    }
+
+    private void OnCommand(string command, string args)
+    {
+        DrawConfigUI();
+    }
+
+    private void DrawUI()
+    {
+        WindowSystem.Draw();
+    }
+
+    public void DrawConfigUI()
+    {
+        ConfigWindow.IsOpen = !ConfigWindow.IsOpen;
     }
 }
