+++ Plugin/PuppetMaster_Enhanced.csproj
@@ -1,11 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Dalamud.NET.Sdk/13.0.0">
+<Project Sdk="Dalamud.NET.Sdk/13.1.0">
   <PropertyGroup>
-    <Authors></Authors>
+    <Authors>Aspher0</Authors>
     <Company></Company>
-    <Version>7.3.1.0</Version>
-    <Description>Let people control your character through chat. A.S. Fork.</Description>
-    <Copyright></Copyright>
+    <Version>1.0.0.69</Version>
+    <Description>Let people control your character through chat. Enhanced version with Whitelist/Blacklist.</Description>
     <PackageProjectUrl>https://github.com/Aspher0/PuppetMaster_Fork</PackageProjectUrl>
     <PackageLicenseExpression>AGPL-3.0-or-later</PackageLicenseExpression>
     <IsPackable>false</IsPackable>
@@ -18,4 +17,8 @@
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <DebugType>none</DebugType>
   </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="NoireLib" Version="1.3.0.4" />
+  </ItemGroup>
 </Project>
