+++ Plugin/Struct/ChannelSetting.cs
@@ -1,0 +1,12 @@
+using Dalamud.Game.Text;
+
+namespace PuppetMaster_Enhanced;
+
+public class ChannelSetting
+{
+    public XivChatType ChatType { get; set; }
+
+    public bool Enabled { get; set; }
+
+    public string Name { get; set; } = string.Empty;
+}
